requirements/requirements.txt
# Django
django==4.2.2
djangorestframework==3.14.0
django-filter==23.2
django-extensions==3.2.3
django-redis==5.3.0
django-cors-headers==4.1.0
django-rosetta==0.9.9

# Authentication
djangorestframework-simplejwt==5.2.2
PyJWT==2.7.0

# Celery
celery==5.3.1
django-celery-results==2.5.1
django-celery-beat==2.5.0

# Others
gunicorn==20.1.0
psycopg2-binary==2.9.6
model-bakery==1.12.0
redis==4.5.5
colorlog==6.7.0
drf-spectacular==0.26.3
pillow==9.5.0
docker==6.1.3
django_prometheus==2.3.1
prometheus-client==0.17.0
psutil==5.9.5
